[feat] SINR-user map and SINR hisotgram added
I have added a function that shows users on the stage with a colour depending on their SINR value.

Now at the end of the simulation the PDF of the SINR values is displayed as a histogram.

Fixed a small bug when generating in the interf_gNBs_rx_power function, where a vector with a wrong length was created.

I have added that the scatters have a variable height.